requirements.txt
requests
uvicorn[standard]
fastapi_utils
pymongo
python-dateutil
fastapi
pytest